+++ db/migrate/20170718152814_add_authentication_token_to_projects.rb
@@ -1,0 +1,6 @@
+class AddAuthenticationTokenToProjects < ActiveRecord::Migration[5.1]
+  def change
+    add_column :projects, :authentication_token, :string
+    add_index  :projects, :authentication_token, unique: true
+  end
+end